refactor: DomainExtensions 패키지 이동
