+++ android/app/src/main/java/io/coursepick/coursepick/presentation/model/DomainExtensions.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.presentation.extension
+package io.coursepick.coursepick.presentation.model
 
 import androidx.annotation.StringRes
 import io.coursepick.coursepick.R
